--- wagon.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-# Wagon
-class Wagon
-  attr_accessor :type
-
-  def initialize(type)
-    @type = type
-  end
-end
